Move weekdays mapping func to utils
